requirements.txt
# Sysdig Report Studio dependencies
streamlit>=1.28.0
pandas>=2.0.0
pyyaml>=6.0
requests>=2.28.0
plotly>=5.18.0
kaleido>=0.2.1
reportlab>=4.0.0
